[SWAGGER] Status TESTED

+++ controller/Roles/RolesController.js
@@ -4,11 +4,16 @@ const roleTable = db['Role'];
 const getRole = async (req, res) => {
     try {
         const role = await roleTable.findByPk(req.params.id);
-
-        res.status(200).send({
-            message: `Role ${role.id}`,
-            data: role
-        })
+        if (role !== null) {
+            res.status(200).send({
+                message: `Role ${role.id}`,
+                data: role
+            })
+        } else {
+            res.status(404).send({
+                message: 'Role not found'
+            })
+        }
     } catch (error) {
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
@@ -20,12 +25,15 @@ const getRole = async (req, res) => {
 const getAllRoles = async (req, res) => {
     try {
         const roles = await roleTable.findAll();
-
-        res.status(200).send({
-            message: 'select all',
-            data: roles
-        })
-
+        if (roles !== null) {
+            res.status(200).send({
+                data: roles
+            })
+        } else {
+            res.status(404).send({
+                message: 'No role found'
+            })
+        }
     } catch (error) {
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
@@ -38,11 +46,16 @@ const createRole = async (req, res) => {
     try {
         let data = { ...req.body };
         const newRole = await roleTable.create(data);
-
-        res.status(200).send({
-            message: 'Role créé',
-            data: newRole
-        })
+        if (newRole !== null) {
+            res.status(200).send({
+                message: 'Role created',
+                data: newRole
+            })
+        } else {
+            res.status(404).send({
+                message: 'Role not created'
+            })
+        }
     } catch (error) {
         res.status(400).send({
             message: 'Erreur de synthaxe de la requête.',
@@ -55,16 +68,27 @@ const modifyRole = async (req, res) => {
     try {
         const { name } = req.body;
         const RoleId = req.params.id;
-        const updateRole = await roleTable.update(
-            { name: name },
-            {
-                where: {
-                    id: RoleId
-                }
-            })
-        if (updateRole[0] == 1) {
-            res.status(200).send({
-                message: 'Role modifié'
+        const idRoleFound = await roleTable.findByPk(RoleId);
+        if (idRoleFound !== null) {
+            const updateRole = await roleTable.update(
+                { name: name },
+                {
+                    where: {
+                        id: RoleId
+                    }
+                })
+            if (updateRole[0] == 1) {
+                res.status(200).send({
+                    message: 'Role updated'
+                })
+            } else {
+                res.status(400).send({
+                    message: 'Role was not updated'
+                })
+            }
+        } else {
+            res.status(404).send({
+                message: 'Role not found'
             })
         }
     } catch (error) {
@@ -78,18 +102,24 @@ const modifyRole = async (req, res) => {
 const deleteRole = async (req, res) => {
     try {
         const RoleId = req.params.id;
-        const deletedRole = await roleTable.destroy({
-            where: { id: RoleId }
-        })
-        console.log(deletedRole);
-        if (deletedRole == 1) {
-            res.status(200).send({
-                message: 'Role supprimé'
+        const idRoleFound = await roleTable.findByPk(RoleId);
+        if (idRoleFound !== null) {
+            const deletedRole = await roleTable.destroy({
+                where: { id: RoleId }
             })
-        }
-        else {
+            if (deletedRole == 1) {
+                res.status(200).send({
+                    message: 'Role deleted'
+                })
+            }
+            else {
+                res.status(400).send({
+                    message: 'Role not deleted'
+                })
+            }
+        } else {
             res.status(404).send({
-                message: '(DELETE) le role n\' a pas été supprimé'
+                message: 'Role not found'
             })
         }
     } catch (error) {
